refactoring to use jwt token/db rename to fit PostgreSQL

+++ src/main/java/com/example/sales/service/DiscountService.java
@@ -1,12 +1,11 @@
 package com.example.sales.service;
 
-import com.example.sales.Enum.RolesEnum;
+import com.example.sales.auth.service.TokenService;
 import com.example.sales.dto.request.DiscountRequestDTO;
 import com.example.sales.dto.request.DiscountUpdateRequestDTO;
 import com.example.sales.dto.response.DiscountResponseDTO;
 import com.example.sales.mapper.DiscountMapper;
 import com.example.sales.model.DiscountEntity;
-import com.example.sales.model.UserEntity;
 import com.example.sales.repository.DiscountRepository;
 import com.example.sales.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -27,18 +26,22 @@ public class DiscountService {
     @Autowired
     private DiscountMapper discountMapper;
 
-    public void createDiscount(DiscountRequestDTO request) throws Exception { //TODO adicionar validacao de usuario com jwt
-        validateUser(request.getUserId());
-        if (!discountRepository.existsByNameAndUserId(request.getName(), request.getUserId())) {
+    @Autowired
+    private TokenService tokenService;
+
+    public void createDiscount(String authorization, DiscountRequestDTO request) throws Exception {
+        Long userId = tokenService.decodeToken(authorization).getClaim("userid").asLong();
+        if (!discountRepository.existsByNameAndUserId(request.getName(), userId)) {
             DiscountEntity discount = discountMapper.discountRequestDTOtoEntity(request);
+            discount.setUserId(userId);
             discountRepository.save(discount);
         } else {
             throw new Exception("This discount is already registered! Please verify on ");
         }
     }
 
-    public List<DiscountResponseDTO> getAllDiscountsByUser(Long userId) throws Exception {
-        validateUser(userId);
+    public List<DiscountResponseDTO> getAllDiscountsByUser(String authorization) throws Exception {
+        Long userId = tokenService.decodeToken(authorization).getClaim("userid").asLong();
         if (userRepository.existsById(userId)) {
             List<DiscountEntity> discountList = discountRepository.findAllByUserId(userId);
             return discountList.stream()
@@ -49,27 +52,28 @@ public class DiscountService {
         }
     }
 
-    public void deleteDiscountById(Long discountId) {
-        DiscountEntity discount = discountRepository.findById(discountId).orElseThrow(() -> new EntityNotFoundException("Discount Not Found!"));
-        discountRepository.delete(discount);
-    }
-
-    //TODO Corrigir: Quando cadastramos um disconto igual, so que para outro usuario ele tbm esta bloqueando. (Isso vai ser corrigido quando tivermos o userId no jwt)
-    public void updateDiscountById(DiscountUpdateRequestDTO request, Long discountId) {
-        DiscountEntity discount = discountRepository.findById(discountId).orElseThrow(() -> new EntityNotFoundException("Discount Not Found!"));
-        discountMapper.updateRequestDTOtoEntity(request, discount);
-        discountRepository.save(discount);
-    }
-
-    public void deleteAllDiscounts(Long userId) throws Exception {
-        validateUser(userId);
-        discountRepository.deleteAllByUserId(userId);
-    }
-
-    public void validateUser(Long userId) throws Exception {
-        UserEntity user = userRepository.findById(userId).orElseThrow(() -> new Exception("User Not Found!"));
-        if (!user.getRole().equals(RolesEnum.STORE)) {
-            throw new Exception("This user can't create a discount");
+    public void deleteDiscountById(String authorization, Long discountId) {
+        DiscountEntity discount = discountRepository.findByIdAndUserId(tokenService.decodeToken(authorization).getClaim("userid").asLong(), discountId);
+        if (discount != null) {
+            discountRepository.delete(discount);
+        } else {
+            throw new EntityNotFoundException("The discount doesn't exists!");
         }
+    }
+
+    public void updateDiscountById(String authorization, DiscountUpdateRequestDTO request, Long discountId) {
+        Long userId = tokenService.decodeToken(authorization).getClaim("userid").asLong();
+        DiscountEntity discount = discountRepository.findByIdAndUserId(userId, discountId);
+        if (discount != null) {
+            discountMapper.updateRequestDTOtoEntity(request, discount);
+            discount.setUserId(userId);
+            discountRepository.save(discount);
+        } else {
+            throw new EntityNotFoundException("Discount Not Found!");
+        }
+    }
+
+    public void deleteAllDiscounts(Long userId) {
+        discountRepository.deleteAllByUserId(userId);
     }
 }
